fix(ui): render mermaid fences in the .md file preview, not just chat
FileViewer's ReactMarkdown instance had no `pre` override, so a ```mermaid
fence in a workspace .md file fell through to a plain <pre><code> block —
literal source text, never a diagram — even though AssistantMessage already
routes the identical fence to the Mermaid renderer for chat messages.

Moved the fence-routing pre-component (mermaidCode + the pre override) out of
AssistantMessage.tsx and into Mermaid.tsx as a shared MarkdownPre export, then
wired it into FileViewer's markdown components too. Same primitive, reused —
per the components spec's SharedPresentationPrimitives requirement — rather
than a second implementation of the same routing.

Verified in the running app: seeded a workspace .md file with a mermaid fence,
served it through a real server + built web UI, and opened it in the file
tree. Before the fix the preview showed literal mermaid source in a <pre>;
after, it renders the actual flowchart SVG.
